fix frontend trip details display

+++ trip_service/src/main/java/pl/edu/pg/trip/enity/Trip.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class Trip {
-    private Long tripId;
+    private Integer tripId;
     private Long hotelId;
     private Long startFlightId;
     private Long endFlightId;
